pyproject.toml
[tool.poetry]
name = "quantum-computing-playground"
version = "0.1.0"
description = "Playground for learning quantum computing using IBM's Qiskit"
authors = ["Zach Bloss <zacharybloss@gmail.com>"]
readme = "README.md"
packages = [{include = "quantum_computing_playground"}]

[tool.poetry.dependencies]
python = "^3.11"
qiskit = {extras = ["visualization"], version = "^0.44.1"}
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
qiskit-ibm-runtime = "^0.12.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
